pyproject.toml
[project]
name = "quantity-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.96.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pydantic>=2.13.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "streamlit>=1.56.0",
    "xlrd>=2.0.2",
]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
